refactor(services): move language service into own module
